Initial steps on conference registration.
This does mostly flight stuff.  Other things will need further changes
to the DB before it will work.

However, it does compile, and the route just always rejects requests.
Rejections even occur for the right reason if the input is invalid,
but will fail saying the dependents are unowned if actually valid
input is given.