Permissions stuff for conference creation/delete.
We now do auth checks for conference creation, as well as only allowing
conference managers to delete a given conference.

This also includes seeder changes to create a user with the new permissions,
so that tests can still pass.  Should probably add tests for when the
user does not have appropriate permissions.